# backend/nestjs/identity-billpay-admin-service/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/health/check":{"post":{"operationId":"HealthController_check","summary":"Health check","description":"Verifies the API and MySQL database are reachable. No authentication required.","parameters":[],"responses":{"200":{"description":"Service is healthy"},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}},"/api/v1/auth/registration/list":{"post":{"operationId":"RegistrationController_list","summary":"[Admin] List registration attempts","description":"**Audience:** Admin web portal (**bank staff** / superadmin). Use Keycloak client `admin-web` for login.\n\nReturns all active customer registration records for bank staff review in the admin portal.","parameters":[],"responses":{"200":{"description":"List of registration attempts"}},"tags":["Mobile — Registration (Customer onboarding)","Auth — Registration"],"security":[{"bearer":[]}]}},"/api/v1/auth/registration/get":{"post":{"operationId":"RegistrationController_get","summary":"[Admin] Get registration attempt by ID","description":"**Audience:** Admin web portal (**bank staff** / superadmin). Use Keycloak client `admin-web` for login.\n\nFetches a single customer registration record by internal UUID for admin review.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"200":{"description":"Registration record"},"404":{"description":"Not found"}},"tags":["Mobile — Registration (Customer onboarding)","Auth — Registration"],"security":[{"bearer":[]}]}},"/api/v1/auth/registration/create":{"post":{"operationId":"RegistrationController_create","summary":"[Mobile / Customer] Start customer registration","description":"**Audience:** Mobile app (retail/corporate **customer**). Use Keycloak client `mobile-app` for login.\n\nFirst step of mobile-app customer onboarding. Creates a registration attempt with step INIT using the customer’s mobile number and PAN/CIF. Public — no Bearer token required.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InitRegistrationDto"}}}},"responses":{"201":{"description":"Registration started"}},"tags":["Mobile — Registration (Customer onboarding)","Auth — Registration"]}},"/api/v1/auth/registration/delete":{"post":{"operationId":"RegistrationController_delete","summary":"[Admin] Soft-delete registration attempt","description":"**Audience:** Admin web portal (**bank staff** / superadmin). Use Keycloak client `admin-web` for login.\n\nMarks a customer registration as deleted for admin cleanup. Record is retained in the database.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"200":{"description":"Soft-deleted successfully"}},"tags":["Mobile — Registration (Customer onboarding)","Auth — Registration"],"security":[{"bearer":[]}]}},"/api/v1/auth/credential/list":{"post":{"operationId":"CredentialController_list","summary":"[Admin] List all customer MPIN credentials","description":"**Audience:** Admin web portal (**bank staff** / superadmin). Use Keycloak client `admin-web` for login.\n\nReturns all active local MPIN records for support and audit. Customers do not use this from the mobile app.","parameters":[],"responses":{"201":{"description":""}},"tags":["Mobile — Credential (Customer MPIN)"],"security":[{"bearer":[]}]}},"/api/v1/auth/credential/get":{"post":{"operationId":"CredentialController_get","summary":"[Mobile / Customer] Get customer MPIN credential","description":"**Audience:** Mobile app (retail/corporate **customer**). Use Keycloak client `mobile-app` for login.\n\nReturns the logged-in customer’s MPIN metadata from the mobile app. Uses the Keycloak user ID (`sub` from POST /auth/session/me). If `userId` is omitted, the authenticated Bearer token user is used.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCredentialDto"}}}},"responses":{"404":{"description":"No MPIN credential exists for this customer"}},"tags":["Mobile — Credential (Customer MPIN)"],"security":[{"bearer":[]}]}},"/api/v1/auth/credential/create":{"post":{"operationId":"CredentialController_create","summary":"[Mobile / Customer] Set customer MPIN (and optional password)","description":"**Audience:** Mobile app (retail/corporate **customer**). Use Keycloak client `mobile-app` for login.\n\nPrimary mobile-app onboarding endpoint. **MPIN** is hashed and stored locally for banking transactions in the mobile app. **Password** (optional) is written to Keycloak only. Use the customer’s Keycloak user ID (`sub` from session/me) as `userId`.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetCredentialDto"}}}},"responses":{"201":{"description":""}},"tags":["Mobile — Credential (Customer MPIN)"],"security":[{"bearer":[]}]}},"/api/v1/auth/credential/delete":{"post":{"operationId":"CredentialController_delete","summary":"[Mobile / Customer] Reset customer MPIN","description":"**Audience:** Mobile app (retail/corporate **customer**). Use Keycloak client `mobile-app` for login.\n\nSoft-deletes the customer’s local MPIN record (e.g. mobile app “forgot MPIN” flow). Does not change the Keycloak login password.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCredentialDto"}}}},"responses":{"201":{"description":""}},"tags":["Mobile — Credential (Customer MPIN)"],"security":[{"bearer":[]}]}},"/api/v1/auth/device/list":{"post":{"operationId":"DeviceController_list","summary":"List all device profiles","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth — Device"],"security":[{"bearer":[]}]}},"/api/v1/auth/device/get":{"post":{"operationId":"DeviceController_get","summary":"Get device profile by ID","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Device"],"security":[{"bearer":[]}]}},"/api/v1/auth/device/create":{"post":{"operationId":"DeviceController_create","summary":"Register a device","description":"Links a mobile device to a user for trusted-device checks.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDeviceDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Device"],"security":[{"bearer":[]}]}},"/api/v1/auth/device/delete":{"post":{"operationId":"DeviceController_delete","summary":"Soft-delete device profile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Device"],"security":[{"bearer":[]}]}},"/api/v1/auth/corporate-hierarchy/list":{"post":{"operationId":"CorporateHierarchyController_list","summary":"List corporate hierarchy records","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth — Corporate Hierarchy"],"security":[{"bearer":[]}]}},"/api/v1/auth/corporate-hierarchy/get":{"post":{"operationId":"CorporateHierarchyController_get","summary":"Get hierarchy record by ID","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Corporate Hierarchy"],"security":[{"bearer":[]}]}},"/api/v1/auth/corporate-hierarchy/create":{"post":{"operationId":"CorporateHierarchyController_create","summary":"Assign corporate role","description":"Links a user to a CIF with a realm role (e.g. CORPORATE_MAKER, CORPORATE_CHECKER).","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetHierarchyRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Corporate Hierarchy"],"security":[{"bearer":[]}]}},"/api/v1/auth/corporate-hierarchy/delete":{"post":{"operationId":"CorporateHierarchyController_delete","summary":"Soft-delete hierarchy record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — Corporate Hierarchy"],"security":[{"bearer":[]}]}},"/api/v1/auth/otp/list":{"post":{"operationId":"OtpController_list","summary":"List all OTP challenges","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth — OTP"],"security":[{"bearer":[]}]}},"/api/v1/auth/otp/get":{"post":{"operationId":"OtpController_get","summary":"Get OTP challenge by ID","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — OTP"],"security":[{"bearer":[]}]}},"/api/v1/auth/otp/create":{"post":{"operationId":"OtpController_create","summary":"Generate OTP","description":"Creates a 6-digit OTP for the given mobile number. Public — no token required. Response includes the OTP (dev only; remove in production).","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateOtpDto"}}}},"responses":{"201":{"description":"OTP generated"}},"tags":["Auth — OTP"]}},"/api/v1/auth/otp/delete":{"post":{"operationId":"OtpController_delete","summary":"Soft-delete OTP challenge","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth — OTP"],"security":[{"bearer":[]}]}},"/api/v1/auth/login":{"post":{"operationId":"SessionController_login","summary":"Login","description":"Authenticates a user against Keycloak (bharat-banking realm) using username/password. Returns access and refresh tokens. Use accessToken as Bearer token for protected APIs.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}},"401":{"description":"Invalid credentials or Keycloak error"}},"tags":["Auth — Session"]}},"/api/v1/auth/logout":{"post":{"operationId":"SessionController_logout","summary":"Logout","description":"Ends the Keycloak session by revoking the refresh token returned from login. Pass the same clientId used during login.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutDto"}}}},"responses":{"200":{"description":"Logout successful"},"401":{"description":"Invalid or expired refresh token"}},"tags":["Auth — Session"]}},"/api/v1/auth/session/me":{"post":{"operationId":"SessionController_me","summary":"Current user profile","description":"Returns the decoded JWT claims for the authenticated user. Requires a valid Bearer access token from login.","parameters":[],"responses":{"200":{"description":"User claims from JWT"},"401":{"description":"Missing or invalid token"}},"tags":["Auth — Session"],"security":[{"bearer":[]}]}},"/api/v1/roles/create":{"post":{"operationId":"RolesController_create","summary":"Create role","description":"Validates the caller Keycloak token for superadmin status, creates the role in Keycloak, and persists it locally.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"Role created and synchronized with Keycloak"}},"tags":["RBAC — Roles"],"security":[{"bearer":[]}]}},"/api/v1/roles/update":{"post":{"operationId":"RolesController_update","summary":"Update role definition","description":"Updates role metadata (display name, description, delegations) in the local database and Keycloak. This does not change an employee assignment — use POST /employees/update-role for that.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["RBAC — Roles"],"security":[{"bearer":[]}]}},"/api/v1/roles/map-permissions":{"post":{"operationId":"RolesController_mapPermissions","summary":"Map permissions to role","description":"Replaces permission mappings for the target role ID with the provided permission IDs.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapRolePermissionsDto"}}}},"responses":{"201":{"description":""}},"tags":["RBAC — Roles"],"security":[{"bearer":[]}]}},"/api/v1/permissions/create":{"post":{"operationId":"PermissionsController_create","summary":"Create permission","description":"Validates superadmin privileges via Keycloak and inserts a new permission record.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePermissionDto"}}}},"responses":{"201":{"description":"Permission created"}},"tags":["RBAC — Permissions"],"security":[{"bearer":[]}]}},"/api/v1/employees/create":{"post":{"operationId":"EmployeesController_create","summary":"Create employee","description":"Provisions a user in Keycloak and local storage. Superadmins can assign any role below their level; bank admins can assign only delegated roles that are below their hierarchy level.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEmployeeDto"}}}},"responses":{"201":{"description":"Employee created in Keycloak and local DB"}},"tags":["RBAC — Employees"],"security":[{"bearer":[]}]}},"/api/v1/employees/update-role":{"post":{"operationId":"EmployeesController_updateRole","summary":"Update employee role","description":"Changes the role assigned to an existing employee in both Keycloak and the local database. You cannot modify employees at your hierarchy level or above, and you cannot assign roles at your level or above. Superadmins can manage all employees; bank admins can manage only lower-hierarchy employees with delegated roles.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEmployeeRoleDto"}}}},"responses":{"201":{"description":"Employee role updated"}},"tags":["RBAC — Employees"],"security":[{"bearer":[]}]}},"/api/v1/users/fetch-access-details":{"post":{"operationId":"UsersController_fetchAccessDetails","summary":"Fetch user access details","description":"Queries Keycloak and the local RBAC tables to return resolved user profiles with roles and permissions.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FetchAccessDetailsDto"}}}},"responses":{"200":{"description":"Resolved access details"}},"tags":["RBAC — Users"],"security":[{"bearer":[]}]}}},"info":{"title":"NPST BCB — Auth Service","description":"Bharat Banking authentication APIs. All endpoints use POST.\n\n**Audiences:**\n- `[Mobile / Customer]` — mobile app endpoints for retail/corporate customers (use Keycloak client `mobile-app`)\n- `[Admin]` — admin web portal for bank staff (use Keycloak client `admin-web`)\n\n**Auth flow:**\n1. `POST /auth/login` — copy **`accessToken`** (not `refreshToken`)\n2. Click **Authorize** and paste the token only (no `Bearer` prefix)\n3. Call protected endpoints within 5 minutes\n4. `POST /auth/logout` when done","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Paste **accessToken** from POST /auth/login only. Do not paste refreshToken or the word Bearer."}},"schemas":{"IdRequestDto":{"type":"object","properties":{"id":{"type":"string","description":"Internal record UUID returned from a list/create response"}},"required":["id"]},"InitRegistrationDto":{"type":"object","properties":{"mobileNumber":{"type":"string","description":"Customer mobile number starting registration in the mobile app","example":"9876543210"},"panOrCif":{"type":"string","description":"Customer PAN or CIF used to identify the corporate/retail customer during mobile onboarding","example":"CIF12345"}},"required":["mobileNumber","panOrCif"]},"GetCredentialDto":{"type":"object","properties":{"userId":{"type":"string","description":"Customer Keycloak user ID (`sub` from session/me). Omit to use the logged-in mobile app user.","example":"334b032c-7468-47fa-82a3-8204b80913a2"}}},"SetCredentialDto":{"type":"object","properties":{"userId":{"type":"string","description":"Keycloak user ID of the **customer** (`sub` from POST /auth/session/me on mobile app)","example":"334b032c-7468-47fa-82a3-8204b80913a2"},"password":{"type":"string","description":"Customer login password for the mobile app. Stored in Keycloak only (never in local DB).","example":"MySecurePass@123"},"mpin":{"type":"string","description":"Customer banking MPIN (4–6 digits) used in the **mobile app**. Hashed and stored locally.","example":"1234"}},"required":["userId"]},"RegisterDeviceDto":{"type":"object","properties":{"userId":{"type":"string","description":"User ID the device belongs to","example":"a1b2c3d4-e5f6-7890-abcd-ef1234567890"},"deviceId":{"type":"string","description":"Unique device identifier (IMEI, UUID, etc.)","example":"device-abc-12345"},"deviceModel":{"type":"string","description":"Human-readable device model name","example":"Samsung Galaxy S24"}},"required":["userId","deviceId"]},"SetHierarchyRoleDto":{"type":"object","properties":{"userId":{"type":"string","description":"Keycloak user ID","example":"a1b2c3d4-e5f6-7890-abcd-ef1234567890"},"cif":{"type":"string","description":"Corporate CIF number","example":"CIF12345"},"role":{"type":"string","description":"Realm role assigned to the user in corporate hierarchy","enum":["BANK_SUPER_ADMIN","BANK_ADMIN","BANK_MAKER","BANK_CHECKER","CORPORATE_IT_ADMIN","CORPORATE_MAKER","CORPORATE_CHECKER","CORPORATE_VIEWER","RETAIL_CUSTOMER"],"example":"CORPORATE_MAKER"}},"required":["userId","cif","role"]},"GenerateOtpDto":{"type":"object","properties":{"mobileNumber":{"type":"string","description":"Mobile number to send the OTP to","example":"9876543210"}},"required":["mobileNumber"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string","description":"Keycloak username / login ID (not email)","example":"corp-maker-01"},"password":{"type":"string","description":"User password","example":"corp-maker-01"},"clientId":{"type":"string","description":"Keycloak client: `mobile-app` for **customer mobile app**, `admin-web` for **bank admin portal**","enum":["admin-web","mobile-app"],"default":"admin-web","example":"mobile-app"}},"required":["username","password"]},"TokenResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"Use this value in Swagger Authorize. RS256 JWT (3 dot-separated parts).","example":"eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUI..."},"expiresIn":{"type":"number","example":300,"description":"Access token lifetime in seconds"},"refreshExpiresIn":{"type":"number","example":1800},"refreshToken":{"type":"string","description":"For logout only. Do NOT use this in Authorize for API calls.","example":"eyJhbGciOiJIUzI1NiIsInR5cCIgOiAiSldUI..."},"tokenType":{"type":"string","example":"Bearer"},"scope":{"type":"string","example":"openid profile email"}},"required":["accessToken","expiresIn","refreshExpiresIn","refreshToken","tokenType","scope"]},"LogoutDto":{"type":"object","properties":{"refreshToken":{"type":"string","description":"Refresh token returned from the login response","example":"eyJhbGciOiJIUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJhYmMxMjMifQ..."},"clientId":{"type":"string","description":"Keycloak client used during login","enum":["admin-web","mobile-app"],"default":"admin-web","example":"admin-web"}},"required":["refreshToken"]},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string","example":"BANK_MAKER"},"displayName":{"type":"string","example":"Bank Maker"},"description":{"type":"string","example":"Creates banking transactions for approval"},"delegatedAdminKeycloakUserIds":{"description":"Keycloak user IDs of BANK_ADMIN users allowed to assign this role","type":"array","items":{"type":"string"}}},"required":["name","displayName"]},"UpdateRoleDto":{"type":"object","properties":{"roleId":{"type":"string","description":"Local role UUID to update"},"displayName":{"type":"string","example":"Bank Maker (Updated)"},"description":{"type":"string","example":"Updated role description"},"isActive":{"type":"boolean","example":true},"delegatedAdminKeycloakUserIds":{"description":"Replace delegated BANK_ADMIN user list for this role","type":"array","items":{"type":"string"}}},"required":["roleId"]},"MapRolePermissionsDto":{"type":"object","properties":{"roleId":{"type":"string","description":"Target local role UUID"},"permissionIds":{"description":"Permission UUIDs to assign to the role (replaces existing mappings)","example":["c12069f4-ad09-41e8-81e2-863c45728307"],"type":"array","items":{"type":"string"}}},"required":["roleId","permissionIds"]},"CreatePermissionDto":{"type":"object","properties":{"code":{"type":"string","example":"EMPLOYEE_CREATE"},"name":{"type":"string","example":"Create Employee"},"description":{"type":"string","example":"Allows provisioning employees in Keycloak and local DB"},"module":{"type":"string","example":"EMPLOYEE"},"action":{"type":"string","example":"CREATE"}},"required":["code","name","module","action"]},"CreateEmployeeDto":{"type":"object","properties":{"username":{"type":"string","example":"bank.maker.02"},"email":{"type":"string","example":"bank.maker.02@bank.example.com"},"firstName":{"type":"string","example":"Ravi"},"lastName":{"type":"string","example":"Sharma"},"password":{"type":"string","example":"BankMaker@123","minLength":8},"roleId":{"type":"string","description":"Local role UUID to assign","example":"c12069f4-ad09-41e8-81e2-863c45728307"},"employeeCode":{"type":"string","example":"EMP-00042"}},"required":["username","email","firstName","lastName","password","roleId"]},"UpdateEmployeeRoleDto":{"type":"object","properties":{"employeeId":{"type":"string","description":"Local employee UUID whose role should be updated","example":"c12069f4-ad09-41e8-81e2-863c45728307"},"roleId":{"type":"string","description":"New local role UUID to assign to the employee","example":"d23070f5-be10-52f9-92f3-974d56839418"}},"required":["employeeId","roleId"]},"FetchAccessDetailsDto":{"type":"object","properties":{"keycloakUserId":{"type":"string","description":"Filter by Keycloak user UUID"},"employeeId":{"type":"string","description":"Filter by local employee UUID"},"username":{"type":"string","description":"Filter by username (exact match)"},"email":{"type":"string","description":"Filter by email (exact match)"},"roleId":{"type":"string","description":"Filter by local role UUID"}}}}}}